refactor: isolate condition policies with explicit environment inputs
